7/25 improved CBox functionality - can now use keyboard to change selected item profile box in the main menu doesn't allow invalid input, such as numbers and other random characters that the bitmap font doesn't support.
Also fixed a problem where the backspace key would add a weird character when you pressed it and there were no characters in the string...which in turn caused it to change the save file name, which made the loading game not work.

CBox also has screen centering and text centering functionaliry.