major overhaul to shredder's AI, everything works well. Only bug is when he tries to move to a turtle and there are obstacles that don't allow the path found to be executed with the amount of AP that he has.
skills have been reorganized. now when the QTE displays, the attack animations WAIT for input, if successful the attack executes along with the combat text.

fixed a bug in the CTimer. wasn't actually stopping the timer when the given end time was reached.

when a weapon is picked up, a message displays as to what you picked up..for a few seconds